dt1 editor: moved dt1 widget into its own package
